Updating Long Datatypes to Int for tight schema

+++ src/ESFA.DC.ILR.Model.Interface/ILLDDAndHealthProblem.cs
@@ -4,6 +4,6 @@
     {
         int LLDDCat { get; }
 
-        long? PrimaryLLDDNullable { get; }
+        int? PrimaryLLDDNullable { get; }
     }
 }
